+++ yandex/data_source_yandex_kubernetes_node_group.go
@@ -182,6 +182,10 @@ func dataSourceYandexKubernetesNodeGroup() *schema.Resource {
 								},
 							},
 						},
+						"name": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
 					},
 				},
 			},